+++ data/src/config.rs
@@ -40,6 +40,7 @@ use crate::serde::{
     deserialize_f32_positive_float_maybe, deserialize_u8_positive_integer_maybe,
 };
 use crate::server::{ConfigMap as ServerMap, ServerName};
+use crate::shortcut::{Commands, KeyBind};
 use crate::{Theme, environment};
 
 pub mod actions;
@@ -561,6 +562,8 @@ impl Config {
         })
         .map_err(|e| Error::Parse(e.to_string()))?;
 
+        keyboard.validate()?;
+
         let servers = ServerMap::new(
             servers,
             sidebar.order_channels_by,
@@ -815,6 +818,8 @@ pub enum Error {
         "Exactly one of sasl.plain.password, sasl.plain.password_file or sasl.plain.password_command must be set."
     )]
     DuplicateSaslPassword,
+    #[error("Keybind \"{}\" is assigned to multiple actions: {}", keybind.as_config_string(), actions.as_config_string())]
+    KeyBindConflict { keybind: KeyBind, actions: Commands },
     #[error("Config does not exist")]
     ConfigMissing,
 }