+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Added:
 - Config option (`preview.card.description_decode_html`) to decode html strings in preview card descriptions
 - Support displaying a larger version of the card images in-app
 - IRCv3 `no-implicit-names` support
+- Keyboard shortcuts validation (e.g no duplicate key binds)
 
 Fixed: